+++ sevenbridges/meta/resource.py
@@ -66,8 +66,14 @@ class ResourceMeta(type):
 
             # get modified data from the instance
             def modified_data(self):
+                metadata = copy.deepcopy(self._dirty.get('metadata'))
                 difference = _data_diff(self._old, self._data.data)
                 self._dirty.update(difference)
+                if metadata:
+                    # File metadata specific patch, otherwise the diff will
+                    # only return changed parameters even when metadata needs
+                    # to be replaced
+                    self._dirty['metadata'] = metadata
                 return self._dirty
 
             def equals(self, other):
@@ -208,3 +214,11 @@ class Resource(six.with_metaclass(ResourceMeta)):
         :return: Field value or None
         """
         return self._data.data.get(name, None)
+
+    def _set(self, key, value):
+        """
+        Set property value in internal storage
+        :param key: Property name
+        :param value: Property value
+        """
+        self._data.data[key] = value