+++ sevenbridges/models/file.py
@@ -69,6 +69,7 @@ class File(Resource):
     storage = CompoundField(FileStorage, read_only=True)
     metadata = CompoundField(Metadata)
     tags = BasicListField()
+    _secondary_files = BasicListField(name='_secondary_files')
 
     def __str__(self):
         return six.text_type('<File: id={id}>'.format(id=self.id))
@@ -85,6 +86,14 @@ class File(Resource):
 
     def is_folder(self):
         return self.type.lower() == self.FOLDER_TYPE
+
+    @property
+    def secondary_files(self):
+        if self._secondary_files:
+            return [
+                File(api=self._api, **data)
+                for data in self._secondary_files
+            ]
 
     @classmethod
     def query(cls, project=None, names=None, metadata=None, origin=None,
@@ -300,11 +309,12 @@ class File(Resource):
         if silent or bool(modified_data):
             # If metadata is to be set
             if 'metadata' in modified_data:
-                if hasattr(self, '_method'):
+                if hasattr(self, '_overwrite_metadata'):
                     self._api.put(
                         url=self._URL['metadata'].format(id=self.id),
                         data=modified_data['metadata']
                     )
+                    delattr(self, '_overwrite_metadata')
                 else:
                     self._api.patch(
                         url=self._URL['metadata'].format(id=self.id),
@@ -364,12 +374,12 @@ class File(Resource):
         self._old = copy.deepcopy(self._data.data)
 
         # If file.metadata = value was executed
-        # file object will have attribute _method='PUT', which tells us
-        # to force overwrite of metadata on the server. This is metadata
-        # specific. Once we reload the resource we delete the attribute
-        # _method from the instance.
+        # file object will have attribute _overwrite_metadata=True,
+        # which tells us to force overwrite of metadata on the server.
+        # This is metadata specific. Once we reload the resource we delete the
+        # attribute _overwrite_metadata from the instance.
         try:
-            delattr(self, '_method')
+            delattr(self, '_overwrite_metadata')
         except AttributeError:
             pass